Refactor configuration mapping system
Replaced `BaseConfigurationMapperBuilder` and `TableConfigurationMapperBuilder` with a new `ConfigurationMapperBuilder` implementation to simplify the codebase and improve maintainability. Updated `AzureTableConfigurationSource` and `ProviderConfigurationBuilder` to use the new `IConfigurationMapper` interface.

Removed legacy classes and interfaces (`ITableConfigurationMapper`, `BaseConfigurationMapperBuilder`) and cleaned up unused code. Introduced `PrimitiveArrayMapper` and related builders for handling array mappings. Added transitional `IConfigurationMapperOld` for legacy support. Adjusted namespaces and imports to reflect the new structure.
