+++ src/AzureTableConfigurationProvider/Configuration/Mapping/PrimitiveArrayMapper.cs
@@ -1,0 +1,5 @@
+﻿namespace AzureTable.Provider.Configuration.Mapping;
+
+internal sealed class PrimitiveArrayMapper
+{
+}
